### 2026-01-01 b368e3c
add 2 new sections
