+++ paper/paper.bib
@@ -119,3 +119,28 @@
  pages={2825--2830},
  year={2011}
 }
+
+@article{Kumar2019,
+  doi = {10.21105/joss.01143},
+  url = {https://doi.org/10.21105/joss.01143},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {33},
+  pages = {1143},
+  author = {Kumar, Ravin and Carroll, Colin and Hartikainen, Ari and Martin, Osvaldo},
+  title = {ArviZ a unified library for exploratory analysis of Bayesian models in Python},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Capretto_Bambi_A_simple_2022,
+author = {Capretto, Tomas and Piho, Camen and Kumar, Ravin and Westfall, Jacob and Yarkoni, Tal and Martin, Osvaldo A},
+doi = {10.18637/jss.v103.i15},
+journal = {Journal of Statistical Software},
+month = aug,
+number = {103},
+pages = {1--29},
+title = {{Bambi: A simple interface for fitting Bayesian linear models in Python}},
+volume = {15},
+year = {2022}
+}
