+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic API and seamless integration with widely used Python data analysis libraries. It covers the majority of common conjugate likelihood-prior pairs found in statistical literature, making rigorous Bayesian updating, exploration, and visualization accessible for practitioners, educators, and researchers. Comprehensive documentation, interactive examples, and an online Distribution Explorer further support flexible application and learning.
+`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic API and seamless integration with widely used Python data analysis libraries. It implements the conjugate likelihood-prior pairs cataloged in Fink's compendium and Wikipedia's conjugate prior table, making rigorous Bayesian updating, exploration, and  visualization accessible for practitioners, educators, and researchers.  Comprehensive documentation, interactive examples, and an online Distribution  Explorer further support flexible application and learning.
 
 # Statement of need
 
@@ -41,12 +41,16 @@ A prior distribution is conjugate to a likelihood when the posterior remains in
 the same distribution family after observing data[@raiffa1961applied].
 Conjugate priors provide closed-form posterior updates and posterior predictive
 distributions, eliminating the need for numerical
-integration[@fink1997compendium]. Classic examples include the
-Beta-Binomial, Gamma-Poisson, and Normal-Normal families, each of which admits
-compact updates for parameters such as success probability, rate, or
-mean[@murphy2007conjugate]. `conjugate-models` implements the majority of
-conjugate pairs cataloged in the literature and reference
-tables[@fink1997compendium].
+integration[@fink1997compendium]. Because these updates are analytic
+rather than iterative, posterior computation is instantaneous regardless of
+data size—enabling real-time interactive exploration and rapid model
+iteration. Classic examples include the Beta-Binomial, Gamma-Poisson, and
+Normal-Normal families, each of which admits compact updates for parameters
+such as success probability, rate, or mean[@murphy2007conjugate].
+`conjugate-models` implements the conjugate pairs cataloged in Fink's
+compendium[@fink1997compendium] and Wikipedia's conjugate prior table.
+The complete list of supported models is maintained at
+[the online documentation](https://williambdean.github.io/conjugate/models/).
 
 ## Motivation
 
@@ -57,15 +61,18 @@ complex—probabilistic frameworks. `conjugate-models` fills this gap for
 users seeking:
 - A composable API that interoperates smoothly with scientific Python libraries (numpy, pandas, polars, matplotlib, and others)
 - Minimal cognitive overhead and easy expression of classic Bayesian updates
-- Coverage of most conjugate prior-likelihood pairs outlined in statistical literature and on the [Wikipedia Conjugate Prior page](https://en.wikipedia.org/wiki/Conjugate_prior)
-- Interactive and educational resources, including an [online Distribution Explorer](https://williambdean.github.io/conjugate/explorer)
+- Coverage of conjugate prior-likelihood pairs outlined in statistical literature, particularly Fink's compendium[@fink1997compendium] and the [Wikipedia Conjugate Prior page](https://en.wikipedia.org/wiki/Conjugate_prior)
+- Interactive and educational resources, including an [online Distribution Explorer](https://williambdean.github.io/conjugate/explorer) that enables real-time exploration of prior-posterior relationships without requiring local installation
 
 ## Problem Statement
 
 Python lacked a dedicated, user-friendly package making Bayesian conjugate
 inference accessible, idiomatic, and didactically powerful—especially one
 offering robust integration with common data tooling and interactive resources
[…]
