+++ paper/paper.bib
@@ -6,14 +6,18 @@
   address = {Boston, MA}
 }
 
-@misc{miller2005conjugate,
-  author = {Miller, Jeff and others},
-  title = {Earliest Known Uses of Some of the Words of Mathematics: "conjugate prior distributions"},
-  year = {2005},
-  note = {Revision of November 13, 2005},
-  url = {https://jeff560.tripod.com/c.html},
-  urldate = {2005-12-02}
+@article{raschka2018mlxtend,
+  title={MLxtend: Providing machine learning and data science utilities and extensions to Python's scientific computing stack},
+  author={Raschka, Sebastian},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={24},
+  pages={638},
+  year={2018},
+  doi={10.21105/joss.00638}
 }
+
+
 
 @techreport{fink1997compendium,
   author = {Fink, Daniel},
@@ -31,14 +35,7 @@
   url = {https://www.cs.ubc.ca/~murphyk/Papers/bayesGauss.pdf}
 }
 
-@book{liu2014statistical,
-  author = {Liu, Han and Wasserman, Larry},
-  title = {Statistical Machine Learning},
-  year = {2014},
-  publisher = {Carnegie Mellon University},
-  url = {https://www.stat.cmu.edu/~larry/=sml/Bayes.pdf},
-  note = {Lecture notes, p.~314}
-}
+
 
 @article{pymc2023,
   title = {{PyMC}: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
