+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic API and seamless integration with widely used Python data analysis libraries. It implements the conjugate likelihood-prior pairs cataloged in Fink's compendium and Wikipedia's conjugate prior table, making rigorous Bayesian updating, exploration, and  visualization accessible for practitioners, educators, and researchers.  Comprehensive documentation, interactive examples, and an online Distribution  Explorer further support flexible application and learning.
+`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic API and seamless integration with widely used Python data analysis libraries. It implements the conjugate likelihood-prior pairs cataloged in Fink's compendium and Wikipedia's conjugate prior table, making rigorous Bayesian updating, exploration, and visualization accessible for practitioners, educators, and researchers. Comprehensive documentation, interactive examples, and an online Distribution Explorer further support flexible application and learning.
 
 # Statement of need
 
@@ -35,8 +35,6 @@ pandas[@The_pandas_development_team_pandas-dev_pandas_Pandas], and
 polars[@polars2024]. The project also provides live interactive
 documentation and an [online Distribution Explorer](https://williambdean.github.io/conjugate/explorer) for real-time model investigation.
 
-## Conjugate Priors
-
 A prior distribution is conjugate to a likelihood when the posterior remains in
 the same distribution family after observing data[@raiffa1961applied].
 Conjugate priors provide closed-form posterior updates and posterior predictive
@@ -44,27 +42,10 @@ distributions, eliminating the need for numerical
 integration[@fink1997compendium]. Because these updates are analytic
 rather than iterative, posterior computation is instantaneous regardless of
 data size—enabling real-time interactive exploration and rapid model
-iteration. Classic examples include the Beta-Binomial, Gamma-Poisson, and
-Normal-Normal families, each of which admits compact updates for parameters
-such as success probability, rate, or mean[@murphy2007conjugate].
-`conjugate-models` implements the conjugate pairs cataloged in Fink's
+iteration. `conjugate-models` implements the conjugate pairs cataloged in Fink's
 compendium[@fink1997compendium] and Wikipedia's conjugate prior table.
 The complete list of supported models is maintained at
 [the online documentation](https://williambdean.github.io/conjugate/models/).
-
-## Motivation
-
-Many real-world and educational Bayesian inference tasks benefit from the
-simplicity and tractability of conjugate models. However, tooling for these
-models in Python has lagged behind full-featured—and often
-complex—probabilistic frameworks. `conjugate-models` fills this gap for
-users seeking:
-- A composable API that interoperates smoothly with scientific Python libraries (numpy, pandas, polars, matplotlib, and others)
-- Minimal cognitive overhead and easy expression of classic Bayesian updates
-- Coverage of conjugate prior-likelihood pairs outlined in statistical literature, particularly Fink's compendium[@fink1997compendium] and the [Wikipedia Conjugate Prior page](https://en.wikipedia.org/wiki/Conjugate_prior)
-- Interactive and educational resources, including an [online Distribution Explorer](https://williambdean.github.io/conjugate/explorer) that enables real-time exploration of prior-posterior relationships without requiring local installation
-
-## Problem Statement
 
[…]
