### 2026-01-05 5910a7c
add helpers for inputs
