+++ conjugate/helpers.py
@@ -1,0 +1,1078 @@
+"""Helper functions to extract sufficient statistics from observations.
+
+These functions bridge the gap between raw data and the summary statistics
+required by conjugate model functions.
+
+Example:
+    >>> from conjugate.distributions import Gamma
+    >>> from conjugate.models import poisson_gamma
+    >>> from conjugate.helpers import poisson_gamma_inputs
+    >>> observations = [3, 5, 2, 4, 1]
+    >>> prior = Gamma(1, 1)
+    >>> posterior = poisson_gamma(**poisson_gamma_inputs(observations), prior=prior)
+
+For PyTensor compatibility:
+    >>> from conjugate.helpers import poisson_gamma_inputs, pytensor_ops
+    >>> inputs = poisson_gamma_inputs(observations, **pytensor_ops)
+"""
+
+import numpy as np
+from typing import Callable
+
+# =============================================================================
+# Default Operations (NumPy-compatible, handle scalars)
+# =============================================================================
+
+
+def _default_sum(x):
+    """Sum elements. Handles both arrays and scalars."""
+    return np.sum(x)
+
+
+def _default_len(x):
+    """Count elements. Handles both arrays and scalars via atleast_1d."""
+    return np.atleast_1d(x).size
+
+
+def _default_prod(x):
+    """Product of elements. Handles both arrays and scalars."""
+    return np.prod(x)
+
+
+def _default_log(x):
+    """Natural logarithm. Handles both arrays and scalars."""
+    return np.log(x)
+
+
+def _default_max(x):
+    """Maximum value. Handles both arrays and scalars."""
+    return np.max(x)
+
+
+def _default_min(x):
+    """Minimum value. Handles both arrays and scalars."""
+    return np.min(x)
+
+
+def _default_mean(x):
+    """Mean of elements. Handles both arrays and scalars."""
+    return np.mean(x)
+
+
+def _default_cos(x):
+    """Cosine. Handles both arrays and scalars."""
+    return np.cos(x)
+
+
+def _default_sin(x):
+    """Sine. Handles both arrays and scalars."""
+    return np.sin(x)
+
+
+def _default_arctan2(y, x):
+    """Two-argument arctangent. Handles both arrays and scalars."""
+    return np.arctan2(y, x)
+
+
+def _default_maximum(x, y):
+    """Element-wise maximum. Handles both arrays and scalars."""
+    return np.maximum(x, y)
+
+
+def _default_reciprocal(x):
+    """Element-wise reciprocal (1/x). Handles both arrays and scalars."""
+    return 1.0 / x
+
+
+def _default_power(x, exp):
+    """Element-wise power. Handles both arrays and scalars."""
+    return np.power(x, exp)
+
+
+# =============================================================================
+# Operation Bundles
+# =============================================================================
+
+numpy_ops = {
+    "sum_fn": _default_sum,
+    "len_fn": _default_len,
+    "prod_fn": _default_prod,
+    "log_fn": _default_log,
+    "max_fn": _default_max,
+    "min_fn": _default_min,
+    "mean_fn": _default_mean,
+    "cos_fn": _default_cos,
+    "sin_fn": _default_sin,
+    "arctan2_fn": _default_arctan2,
+    "maximum_fn": _default_maximum,
+    "reciprocal_fn": _default_reciprocal,
+    "power_fn": _default_power,
+    "asarray_fn": np.asarray,
+}
+
+
+class _PyTensorOpsProxy:
+    """Lazy loader for PyTensor operations to avoid hard dependency."""
+
+    _ops = None
+
+    def _get_ops(self):
+        """Lazy import PyTensor operations."""
+        if self._ops is None:
+            try:
+                import pytensor.tensor as pt
+
[…]
