### 2026-01-05 56534f5
add tests for helpers
