+++ tests/test_helpers.py
@@ -1,0 +1,720 @@
+"""Tests for conjugate.helpers module."""
+
+import pytest
+import numpy as np
+
+from conjugate.helpers import (
+    # Simple sum/count
+    poisson_gamma_inputs,
+    exponential_gamma_inputs,
+    gamma_known_shape_inputs,
+    # Success/trial counts
+    binomial_beta_inputs,
+    bernoulli_beta_inputs,
+    negative_binomial_beta_inputs,
+    geometric_beta_inputs,
+    categorical_dirichlet_inputs,
+    multinomial_dirichlet_inputs,
+    # Normal variants
+    normal_inputs,
+    normal_known_mean_inputs,
+    normal_known_variance_inputs,
+    log_normal_inputs,
+    # Products
+    gamma_known_rate_inputs,
+    gamma_inputs,
+    beta_inputs,
+    # Log sums
+    pareto_gamma_inputs,
+    # Max/min
+    uniform_pareto_inputs,
+    # Reciprocals
+    inverse_gamma_known_rate_inputs,
+    # Von Mises
+    von_mises_known_concentration_inputs,
+    von_mises_known_direction_inputs,
+    # Weibull
+    weibull_inverse_gamma_known_shape_inputs,
+    # Multivariate
+    multivariate_normal_known_covariance_inputs,
+    multivariate_normal_known_precision_inputs,
+    # Identity
+    linear_regression_inputs,
+    multivariate_normal_inputs,
+    multivariate_normal_known_mean_inputs,
+    # Operation bundles
+    numpy_ops,
+    pytensor_ops,
+)
+
+
+class TestSimpleSumCount:
+    """Test functions that extract x_total and n."""
+
+    def test_poisson_gamma_inputs_array(self):
+        data = [3, 5, 2, 4, 1]
+        result = poisson_gamma_inputs(data)
+        expected = {"x_total": 15, "n": 5}
+        assert result == expected
+
+    def test_poisson_gamma_inputs_scalar(self):
+        data = 7
+        result = poisson_gamma_inputs(data)
+        expected = {"x_total": 7, "n": 1}
+        assert result == expected
+
+    def test_exponential_gamma_inputs_is_alias(self):
+        """Test that exponential_gamma_inputs is the same as poisson_gamma_inputs."""
+        data = [1.5, 2.3, 0.8]
+        result1 = poisson_gamma_inputs(data)
+        result2 = exponential_gamma_inputs(data)
+        assert result1 == result2
+
+    def test_gamma_known_shape_inputs(self):
+        data = np.array([1.2, 3.4, 2.1, 0.9])
+        result = gamma_known_shape_inputs(data)
+        expected = {"x_total": 7.6, "n": 4}
+        assert result == expected
+
+
+class TestSuccessTrialCounts:
+    """Test functions that handle successes and trials."""
+
+    def test_binomial_beta_inputs(self):
+        successes = [5, 8, 3]
+        trials = [10, 12, 8]
+        result = binomial_beta_inputs(successes, trials)
+        expected = {"x": 16, "n": 30}
+        assert result == expected
+
+    def test_bernoulli_beta_inputs(self):
+        data = [1, 0, 1, 1, 0]
+        result = bernoulli_beta_inputs(data)
+        expected = {"x": 3, "n": 5}
+        assert result == expected
+
+    def test_negative_binomial_beta_inputs(self):
+        successes = [2, 3, 1]
+        failures_per_exp = 5
+        result = negative_binomial_beta_inputs(successes, failures_per_exp)
+        expected = {"x": 6, "n": 3, "r": 5}
+        assert result == expected
+
+    def test_geometric_beta_inputs(self):
+        data = [3, 1, 4, 2]
[…]
