+++ README.md
@@ -18,6 +18,7 @@ pip install conjugate-models
 ## Features
 
 - [Interactive Distribution Explorer](https://williambdean.github.io/conjugate/explorer) for exploring probability distributions with real-time parameter adjustment
+- **Data Input Helper Functions** - Extract sufficient statistics from raw observational data for all supported models
 - [Connection to Scipy Distributions](https://williambdean.github.io/conjugate/examples/scipy-connection) with `dist` attribute
 - [Built in Plotting](https://williambdean.github.io/conjugate/examples/plotting) with `plot_pdf`, `plot_pmf`, and `plot_cdf` methods
 - [Vectorized Operations](https://williambdean.github.io/conjugate/examples/vectorized-inputs) for parameters and data
@@ -38,6 +39,8 @@ Many likelihoods are supported including
 
 ## Basic Usage
 
+### Working with Pre-processed Data
+
 1. Define prior distribution from `distributions` module
 1. Pass data and prior into model from `models` modules
 1. Analytics with posterior and posterior predictive distributions
@@ -46,9 +49,9 @@ Many likelihoods are supported including
 from conjugate.distributions import Beta, BetaBinomial
 from conjugate.models import binomial_beta, binomial_beta_predictive
 
-# Observed Data
-x = 4
-N = 10
+# Observed Data (sufficient statistics)
+x = 4  # successes
+N = 10 # trials
 
 # Analytics
 prior = Beta(1, 1)
@@ -59,6 +62,62 @@ posterior_predictive: BetaBinomial = binomial_beta_predictive(
     n=N, distribution=posterior
 )
 ```
+
+### Working with Raw Observational Data
+
+For raw data, use **helper functions** from the `helpers` module to extract sufficient statistics:
+
+```python
+import numpy as np
+from conjugate.distributions import Beta
+from conjugate.models import binomial_beta
+from conjugate.helpers import bernoulli_beta_inputs
+
+# Raw observational data - individual trial outcomes
+raw_data = [1, 0, 1, 1, 0, 1, 0, 1, 1, 0]  # success/failure per trial
+
+# Extract sufficient statistics automatically
+inputs = bernoulli_beta_inputs(raw_data)
+print(inputs)  # {'x': 6, 'n': 10} - 6 successes in 10 trials
+
+# Use with conjugate model
+prior = Beta(1, 1)
+posterior = binomial_beta(prior=prior, **inputs)
+```
+
+#### Common Helper Function Patterns
+
+```python
+from conjugate.helpers import (
+    poisson_gamma_inputs,      # For count data
+    normal_known_variance_inputs,  # For continuous measurements
+    exponential_gamma_inputs,  # For time-between-events data
+    multinomial_dirichlet_inputs,  # For categorical data
+)
+
+# Count data (e.g., website visits per day)
+count_data = [5, 3, 8, 2, 6, 4, 7, 1, 9, 3]
+inputs = poisson_gamma_inputs(count_data)
+# Returns: {'x': sum(count_data), 'n': len(count_data)}
+
+# Continuous measurements with known variance
+measurements = [2.3, 1.9, 2.7, 2.1, 2.5]
+variance = 0.5
+inputs = normal_known_variance_inputs(measurements, variance=variance)
+# Returns: {'x_mean': mean(measurements), 'n': len(measurements), 'variance': variance}
+
+# Time between events (e.g., customer arrivals)
+wait_times = [3.2, 1.8, 4.1, 2.7, 3.9]
+inputs = exponential_gamma_inputs(wait_times)
+# Returns: {'x': sum(wait_times), 'n': len(wait_times)}
+
+# Categorical outcomes (e.g., survey responses A, B, C)
[…]
