+++ docs/examples/raw-data-workflow.md
@@ -1,0 +1,258 @@
+# Raw Data to Posterior: Complete Workflow
+
+This example demonstrates the complete workflow from raw observational data to Bayesian posterior distributions using the helper functions in `conjugate-models`.
+
+## Overview
+
+Instead of manually computing sufficient statistics, you can use helper functions to extract them automatically from raw observational data. This is particularly useful when working with:
+
+- Experimental data from A/B tests
+- Time series observations
+- Survey responses
+- Measurement data from sensors
+- Count data from web analytics
+
+## Example 1: A/B Testing with Conversion Rates
+
+Suppose you're running an A/B test on a website and want to analyze conversion rates.
+
+```python
+import numpy as np
+import matplotlib.pyplot as plt
+from conjugate.distributions import Beta
+from conjugate.models import binomial_beta
+from conjugate.helpers import bernoulli_beta_inputs
+
+# Raw observational data - user conversion outcomes
+# 1 = converted, 0 = did not convert
+variant_a = [1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1]  # 8 conversions out of 15 users
+variant_b = [1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1]  # 13 conversions out of 20 users
+
+# Extract sufficient statistics automatically
+inputs_a = bernoulli_beta_inputs(variant_a)
+inputs_b = bernoulli_beta_inputs(variant_b)
+
+print("Variant A:", inputs_a)  # {'x': 8, 'n': 15}
+print("Variant B:", inputs_b)  # {'x': 13, 'n': 20}
+
+# Set up priors (Beta(1,1) = uniform prior)
+prior = Beta(1, 1)
+
+# Compute posteriors using extracted statistics
+posterior_a = binomial_beta(**inputs_a, prior=prior)
+posterior_b = binomial_beta(**inputs_b, prior=prior)
+
+print(f"Variant A: Beta({posterior_a.alpha}, {posterior_a.beta})")
+print(f"Variant B: Beta({posterior_b.alpha}, {posterior_b.beta})")
+
+# Plot results
+fig, ax = plt.subplots(figsize=(10, 6))
+x = np.linspace(0, 1, 1000)
+
+ax.plot(x, posterior_a.pdf(x), label=f'Variant A (8/15)', linewidth=2)
+ax.plot(x, posterior_b.pdf(x), label=f'Variant B (13/20)', linewidth=2)
+ax.axvline(inputs_a['x'] / inputs_a['n'], color='blue', alpha=0.5, linestyle='--', label='A MLE')
+ax.axvline(inputs_b['x'] / inputs_b['n'], color='orange', alpha=0.5, linestyle='--', label='B MLE')
+
+ax.set_xlabel('Conversion Rate')
+ax.set_ylabel('Density')
+ax.set_title('A/B Test Results: Posterior Distributions')
+ax.legend()
+ax.grid(alpha=0.3)
+plt.show()
+
+# Statistical comparison
+prob_b_better = (posterior_b.rvs(10000) > posterior_a.rvs(10000)).mean()
+print(f"Probability that B > A: {prob_b_better:.3f}")
+```
+
+## Example 2: Website Analytics - Daily Page Views
+
+Analyzing daily page view counts using Poisson-Gamma conjugacy.
+
+```python
+from conjugate.distributions import Gamma
+from conjugate.models import poisson_gamma, poisson_gamma_predictive
+from conjugate.helpers import poisson_gamma_inputs
+
+# Raw data: daily page views over 30 days
+daily_views = [
[…]

[…]
