### 2026-01-05 3671d52
add scripts
