+++ scripts/generate_workflow_examples.py
@@ -1,0 +1,615 @@
+#!/usr/bin/env python3
+"""
+Generate images for raw-data-workflow.md examples
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from pathlib import Path
+
+# Ensure the images directory exists
+images_dir = Path("docs/images")
+images_dir.mkdir(exist_ok=True)
+
+
+def generate_ab_testing_example():
+    """Generate A/B Testing example and image"""
+    from conjugate.distributions import Beta
+    from conjugate.models import binomial_beta
+    from conjugate.helpers import bernoulli_beta_inputs
+
+    print("=" * 60)
+    print("EXAMPLE 1: A/B Testing with Conversion Rates")
+    print("=" * 60)
+
+    # Raw observational data - user conversion outcomes
+    variant_a = [1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1]  # 8/15 conversions
+    variant_b = [
+        1,
+        1,
+        0,
+        1,
+        1,
+        0,
+        1,
+        0,
+        0,
+        1,
+        1,
+        1,
+        0,
+        1,
+        0,
+        1,
+        1,
+        0,
+        1,
+        1,
+    ]  # 13/20 conversions
+
+    # Extract sufficient statistics automatically
+    inputs_a = bernoulli_beta_inputs(variant_a)
+    inputs_b = bernoulli_beta_inputs(variant_b)
+
+    print("Variant A:", inputs_a)
+    print("Variant B:", inputs_b)
+
+    # Set up priors
+    prior = Beta(1, 1)
+
+    # Compute posteriors
+    posterior_a = binomial_beta(**inputs_a, prior=prior)
+    posterior_b = binomial_beta(**inputs_b, prior=prior)
+
+    print(f"Variant A: Beta({posterior_a.alpha}, {posterior_a.beta})")
+    print(f"Variant B: Beta({posterior_b.alpha}, {posterior_b.beta})")
+
+    # Create visualization
+    fig, ax = plt.subplots(figsize=(10, 6))
+    x = np.linspace(0, 1, 1000)
+
+    ax.plot(
+        x, posterior_a.dist.pdf(x), label="Variant A (8/15)", linewidth=2, color="blue"
+    )
+    ax.plot(
+        x,
+        posterior_b.dist.pdf(x),
+        label="Variant B (13/20)",
+        linewidth=2,
+        color="orange",
+    )
+    ax.axvline(
+        inputs_a["x"] / inputs_a["n"],
+        color="blue",
+        alpha=0.5,
+        linestyle="--",
+        label="A MLE",
+    )
+    ax.axvline(
+        inputs_b["x"] / inputs_b["n"],
+        color="orange",
+        alpha=0.5,
+        linestyle="--",
+        label="B MLE",
+    )
+
+    ax.set_xlabel("Conversion Rate")
+    ax.set_ylabel("Density")
+    ax.set_title("A/B Test Results: Posterior Distributions")
+    ax.legend()
+    ax.grid(alpha=0.3)
+
+    # Save the plot
+    plt.tight_layout()
+    plt.savefig("docs/images/raw-data-ab-testing.png", dpi=150, bbox_inches="tight")
+    plt.close()
+
+    # Statistical comparison
+    prob_b_better = (posterior_b.dist.rvs(10000) > posterior_a.dist.rvs(10000)).mean()
+    print(f"Probability that B > A: {prob_b_better:.3f}")
+    print("✅ Example 1 complete - image saved to docs/images/raw-data-ab-testing.png")
+
+
+def generate_website_analytics_example():
+    """Generate Website Analytics example and image"""
+    from conjugate.distributions import Gamma
[…]
