### 2026-01-06 eb8314a
use python code blocks
