### 2026-01-06 c2ea38a
include images
