### 2026-01-06 801eb94
feat: add Raw Data Workflow feature to documentation
- Add Raw Data Workflow as prominent feature in README.md and docs/index.md
- Include comprehensive docstring formatting guidelines in AGENTS.md
- Emphasize complete workflow from raw observations to posterior distributions
- Link to examples/raw-data-workflow page with 5 real-world case studies
