### 2026-01-06 361187a
docs: update feedback response with current API status
