### 2026-01-06 da975c3
chore: stop tracking prompts/ directory
- Remove prompts/feedback.md from git tracking
- Add prompts/ directory to gitignore
- Keep files local for development reference
