+++ prompts/feedback.md
@@ -1,176 +1,0 @@
-# Improving conjugate-models: Lessons from PyMC Integration - UPDATE
-
-## Overview
-
-While integrating conjugate prior-likelihood relationships into PyMC-Extras' optimization framework, several gaps emerged in the `conjugate-models` package that, if addressed, would make it significantly more useful for downstream consumers.
-
-**UPDATE**: Since this feedback was provided, the `conjugate-models` package has implemented significant improvements addressing the core issues identified below.
-
-## Current State
-
-The `conjugate-models` package provides:
-- ✅ Posterior parameter formulas for ~50+ conjugate pairs
-- ✅ Clean Python API with NumPy/SciPy integration
-- ✅ Distribution objects with analytical methods (mean, variance, pdf, etc.)
-- ✅ **NEW: Complete helper function library for sufficient statistics extraction**
-- ✅ **NEW: Raw data workflow with 50+ helper functions**
-- ✅ **NEW: Interactive documentation with real-world examples**
-
-## Status of Previously Missing Components
-
-### 1. Sufficient Statistics Extraction ✅ **IMPLEMENTED**
-
-**Problem**: Each conjugacy requires specific sufficient statistics from observed data, but this logic isn't stored in the package.
-
-**SOLUTION IMPLEMENTED**: The package now includes a comprehensive `helpers` module with 50+ functions:
-
-```python
-from conjugate.helpers import (
-    poisson_gamma_inputs,      # For count data
-    bernoulli_beta_inputs,     # For binary data
-    normal_inputs,             # For continuous data
-    exponential_gamma_inputs,  # For time-between-events
-    multinomial_dirichlet_inputs,  # For categorical data
-)
-
-# Raw observational data -> sufficient statistics
-count_data = [5, 3, 8, 2, 6, 4, 7, 1, 9, 3]
-inputs = poisson_gamma_inputs(count_data)
-# Returns: {'x_total': 48, 'n': 10}
-
-# Use with conjugate model
-from conjugate.models import poisson_gamma
-from conjugate.distributions import Gamma
-prior = Gamma(alpha=2, beta=1)
-posterior = poisson_gamma(prior=prior, **inputs)
-```
-
-**Complete Coverage**: Helper functions exist for all supported conjugate models:
-- Simple sum/count: `poisson_gamma_inputs`, `exponential_gamma_inputs`
-- Success/trial counts: `binomial_beta_inputs`, `bernoulli_beta_inputs`
-- Sum and sum of squares: `normal_inputs`, `normal_known_mean_inputs`
-- Products and logarithms: `gamma_inputs`, `pareto_gamma_inputs`
-- And many more specialized cases
-
-### 2. Raw Data Workflow Documentation ✅ **IMPLEMENTED**
-
-**NEW FEATURE**: Complete [Raw Data Workflow](https://williambdean.github.io/conjugate/examples/raw-data-workflow) with 5 real-world examples:
-
-1. **A/B Testing** - Beta-Binomial for conversion rate analysis
-2. **Website Analytics** - Poisson-Gamma for daily page view modeling
-3. **Customer Surveys** - Multinomial-Dirichlet for response analysis
-4. **Sensor Data** - Normal-Inverse-Gamma for temperature measurements
-5. **Customer Arrivals** - Exponential-Gamma for time-between-events
-
-Each example shows:
-- Raw observational data input
-- Helper function usage for sufficient statistics
-- Posterior computation
-- Visualization with uncertainty quantification
-
-### 3. Parameterization Documentation/Conversion ⚠️ **PARTIALLY ADDRESSED**
-
-**Status**: The package uses consistent parameterizations and provides clear documentation, but explicit conversion helpers are not yet implemented.
-
-Current state:
[…]
