+++ .github/workflows/docs.yml
@@ -15,6 +15,6 @@ jobs:
           fetch-depth: 0
       - uses: actions/setup-python@v5
         with:
-          python-version: "3.10"
+          python-version: "3.13"
       - run: pip install mkdocs mkdocs-marimo mkdocs-material mkdocs-markdownextradata-plugin mkdocs-git-revision-date-localized-plugin "mkdocstrings[python]"
       - run: mkdocs gh-deploy --force
