+++ .github/workflows/coverage.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Set up Python 3.10
         uses: actions/setup-python@v5
         with:
-          python-version: '3.10'
+          python-version: "3.13"
       - name: Install dependencies
         run: pip install . pytest pytest-cov pytest-mpl pypika
       - name: Run tests and collect coverage
