### 2026-01-06 f0322fe
point out helpers module and add References header
