+++ paper/paper.md
@@ -63,87 +63,29 @@ A typical workflow follows a consistent pattern:
 
 ```python
 from conjugate.distributions import SomeDistribution
+from conjugate.helpers import some_model_inputs
 from conjugate.models import some_model
 
+observed_data = ...
+
 prior: SomeDistribution = ...
-data = ...
-summary_stats = f(data)
-posterior: SomeDistribution = some_model(*summary_stats, prior=prior)
+posterior: SomeDistribution = some_model(
+    **some_model_inputs(observed_data),
+    prior=prior,
+)
 ```
 
 # Example Usage
 
 ## Sequential Bayesian Updates
 
-The package naturally supports sequential Bayesian learning, where each
-posterior becomes the next prior. This example demonstrates real-time updating
-for a binomial model with beta prior:
-
-```python
-from conjugate.distributions import Beta
-from conjugate.models import binomial_beta
-
-# Initial prior: uniform over [0, 1]
-prior = Beta(alpha=1, beta=1)
-
-# Observations: 3 successes in 5 trials, then 2 successes in 3 trials
-observations = [(3, 5), (2, 3)]
-
-posterior = prior
-for successes, trials in observations:
-    posterior = binomial_beta(n=trials, x=successes, prior=posterior)
-    print(f"After {successes}/{trials}: α={posterior.alpha}, β={posterior.beta}")
-
-# After 3/5: α=4.0, β=3.0
-# After 2/3: α=6.0, β=4.0
-```
+The package naturally supports sequential Bayesian learning, where each posterior becomes the next prior. For example, with a binomial model and beta prior, users can iteratively update beliefs as new trial data arrives, with each posterior distribution serving as the prior for the next update. This enables real-time learning applications where beliefs continuously evolve with incoming evidence. A complete worked example of sequential updating with beta-binomial conjugacy is available in the [online documentation](https://williambdean.github.io/conjugate/examples/bayesian-update/).
 
 ![Sequential Bayesian updates showing prior, intermediate posterior, and final posterior distributions. Each update incorporates new evidence while maintaining uncertainty.](bayesian-update.png)
 
 ## Thompson Sampling for Minimizing Wait Times
 
-Thompson sampling is effective for exploration-exploitation problems where the
-goal is optimization. This example demonstrates finding the group with minimum
-wait time using exponential-gamma conjugate updates:
-
-```python
-from conjugate.distributions import Gamma, Exponential
-from conjugate.models import exponential_gamma
-import numpy as np
-
-# Five groups with unknown exponential wait times
-lam = np.array([0.5, 0.55, 0.6, 0.8, 1])  # True rates (hidden)
-n_groups = len(lam)
-true_dist = Exponential(lam=lam)
-
-def thompson_step(estimate: Gamma, rng) -> Gamma:
-    # Sample rate from posterior for each group
-    sample = estimate.dist.rvs(random_state=rng)
-
-    # Choose group with minimum expected wait time (highest rate)
-    group_to_sample = np.argmin(sample)
-
-    # Observe wait time from chosen group
-    group_sample = true_dist[group_to_sample].dist.rvs(random_state=rng)
-
-    # Prepare statistics for Bayesian update
-    x = np.zeros(n_groups)
-    n = np.zeros(n_groups)
-    x[group_to_sample] = group_sample
-    n[group_to_sample] = 1
-
-    return exponential_gamma(x_total=x, n=n, prior=estimate)
-
-# Initialize with uniform priors
-alpha = beta = np.ones(n_groups)
-estimate = Gamma(alpha, beta)
-
-rng = np.random.default_rng(42)
-
-# Thompson sampling over 250 iterations
-for _ in range(250):
-    estimate = thompson_step(estimate=estimate, rng=rng)
-```
[…]
