+++ .pre-commit-config.yaml
@@ -17,3 +17,12 @@ repos:
       - id: end-of-file-fixer
       - id: check-yaml
       - id: check-added-large-files
+  - repo: local
+    hooks:
+      - id: check-docs-frontmatter
+        name: Check docs frontmatter for comments
+        entry: uv run python scripts/check_frontmatter.py
+        language: system
+        files: ^docs/.*\.md$
+        exclude: ^docs/index\.md$
+        pass_filenames: false
