### 2026-01-19 #9704 @matt-graham
@gcanasherrera @aneeshnaik @eigenfoo can you review this submission for JOSS?
