### 2026-01-22 #9704 @matt-graham
@mrazomej @matteodelucchi @limengbinggz can you review this submission for JOSS?
