### 2026-01-23 #9704 @matt-graham
Thanks @GMarupilla for [agreeing to review this submission](https://github.com/openjournals/joss-reviews/issues/9786#issuecomment-3790194929).

We now have three assigned reviewers so I will create the main review issue and close this issue.
