### 2026-01-23 #9890 @editorialbot
[REVIEW]: conjugate-models: Conjugate Models in Python
**Submitting author:** @williambdean (<a href="http://orcid.org/0009-0003-6510-3545">William Dean</a>)
**Repository:** https://github.com/williambdean/conjugate
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.14.2
