### 2026-02-02 #9890 @gmarupilla
@williambdean, I was trying to run some examples from [here](https://williambdean.github.io/conjugate/examples/raw-data-workflow/), but it looks like a few method(s)are not available in the [module](https://github.com/williambdean/conjugate/blob/main/conjugate/models.py#L591)(example: `'Gamma' object has no attribute 'mean'`). I tried with both pip install the package or a full install with the dep group from toml.

Can I open an MR fixing a few examples?
