### 2026-02-02 #9890 @williambdean
Good catch. It should be Gamma(...).dist.mean()
