### 2026-02-10 #9890 @gmarupilla
@williambdean, I’ve completed my review. Overall the package is solid and installation worked locally and the test suite mostly passes. I ran `pytest tests --numprocesses auto`: 218 passed, 3 skipped, and 19 failures were pytest-mpl image baseline mismatches in tests/test_example_plots.py, which appear system‑dependent rather than functional regressions(should be okay). The manuscript is clear and includes all required sections. The one concrete fix I recommend is adding a reference for [Wikipedia’s conjugate prior table](https://github.com/williambdean/conjugate/blob/paper/paper/paper.md#:~:text=Fink%27s%20compendium%20and-,Wikipedia%27s%20conjugate%20prior%20table,-%2C%20making%20rigorous%20Bayesian) which is cited in the text but missing from paper.bib. Here is a suggested BibTeX entry you can use:

```bib
@misc{wikipedia_conjugate_prior,
  title = {Conjugate prior},
  author = {{Wikipedia contributors}},
  year = {2026},
  url = {https://en.wikipedia.org/wiki/Conjugate_prior},
  note = {Accessed 2026-02-09}
}
```
