### 2026-02-10 #9890 @williambdean
Thanks for flagging. I'm working on adding that to the bib.

I'm not too concerned about those tests as I run a matrix across versions and os with consistent passing.
