### 2026-02-10 ea3e2a2
add wiki table to bib
