+++ paper/paper.bib
@@ -131,13 +131,21 @@
 }
 
 @article{Capretto_Bambi_A_simple_2022,
-author = {Capretto, Tomas and Piho, Camen and Kumar, Ravin and Westfall, Jacob and Yarkoni, Tal and Martin, Osvaldo A},
-doi = {10.18637/jss.v103.i15},
-journal = {Journal of Statistical Software},
-month = aug,
-number = {103},
-pages = {1--29},
-title = {{Bambi: A simple interface for fitting Bayesian linear models in Python}},
-volume = {15},
-year = {2022}
+ author = {Capretto, Tomas and Piho, Camen and Kumar, Ravin and Westfall, Jacob and Yarkoni, Tal and Martin, Osvaldo A},
+ doi = {10.18637/jss.v103.i15},
+ journal = {Journal of Statistical Software},
+ month = aug,
+ number = {103},
+ pages = {1--29},
+ title = {{Bambi: A simple interface for fitting Bayesian linear models in Python}},
+ volume = {15},
+ year = {2022}
+}
+
+@misc{wikipedia_conjugate_prior,
+  title = {Conjugate prior},
+  author = {{Wikipedia contributors}},
+  year = {2026},
+  url = {https://en.wikipedia.org/wiki/Conjugate_prior},
+  note = {Accessed 2026-02-09}
 }
