### 2026-02-10 e117e63
tests with plots
