### 2026-02-10 07a316a
chore: test for plots (#304)
* test figures

* add pytest-env

* tests with plots

* move into conftest

* add pytest-env to tox env
