### 2026-02-10 927b9b9
Add frontmatter to Think Bayes examples
Add YAML frontmatter with 'comments: true' to all three new documentation files as required by pre-commit hook.
