+++ docs/examples/think-bayes-euro.md
@@ -1,0 +1,196 @@
+---
+comments: true
+---
+
+# Euro Problem: From Think Bayes Chapter 18
+
+This example ports the Euro problem from Allen Downey's [Think Bayes Chapter 18](https://allendowney.github.io/ThinkBayes2/chap18.html#binomial-likelihood) to demonstrate the Beta-Binomial conjugate relationship.
+
+The problem: estimating the probability of a coin landing heads up, given 140 heads out of 250 flips. The Beta distribution serves as the conjugate prior for the Binomial likelihood, making Bayesian updates trivial.
+
+## Import modules
+
+Import the required distributions and functions:
+
+- `Beta`: Prior and posterior distribution for success probability
+- `Binomial`: Likelihood function for coin flips
+- `BetaBinomial`: Predictive distribution for binomial model
+- `binomial_beta`: Posterior update function
+- `binomial_beta_predictive`: Predictive distribution function
+
+```python
+from conjugate.distributions import Beta, Binomial, BetaBinomial
+from conjugate.models import binomial_beta, binomial_beta_predictive
+
+import matplotlib.pyplot as plt
+import numpy as np
+```
+
+## Observed Data
+
+From Think Bayes: 140 heads out of 250 coin flips. We're estimating the probability `p` of getting heads.
+
+```python
+# Euro coin data: 140 heads out of 250 flips
+x = 140  # number of successes (heads)
+n = 250   # number of trials (flips)
+```
+
+## Bayesian Inference
+
+### Posterior Distribution
+
+Using the Beta-Binomial conjugate relationship, we update our prior with the observed data. The posterior is also a Beta distribution with updated parameters:
+
+```python
+# Prior: Uniform distribution Beta(alpha=1, beta=1) from Think Bayes
+prior = Beta(alpha=1, beta=1)
+
+# Posterior update using conjugate relationship
+posterior = binomial_beta(n=n, x=x, prior=prior)
+
+print(f"Prior parameters: alpha={prior.alpha}, beta={prior.beta}")
+print(f"Posterior parameters: alpha={posterior.alpha}, beta={posterior.beta}")
+print(f"Prior mean: {prior.mean():.3f}")
+print(f"Posterior mean: {posterior.mean():.3f}")
+print(f"Observed proportion: {x/n:.3f}")
+```
+
+The posterior follows the simple conjugate update rule:
+- `α_posterior = α_prior + x` (adding observed heads)
+- `β_posterior = β_prior + (n - x)` (adding observed tails)
+
+### Predictive Distribution
+
+Get the predictive distribution for future coin flips:
+
+```python
+# Predictive distribution for next 10 flips
+n_future = 10
+posterior_predictive = binomial_beta_predictive(
+    n=n_future,
+    distribution=posterior
+)
+
+print(f"Predicted heads in {n_future} flips: {posterior_predictive.mean():.2f}")
+```
+
+## Additional Analysis
+
+Compare prior, posterior, and predictive distributions:
+
+```python
+# Create visualization
+fig, axes = plt.subplots(2, 1, figsize=(10, 8))
+
+# Top plot: Prior vs Posterior distributions
+ax1 = axes[0]
+p_range = np.linspace(0, 1, 200)
+
+# Plot prior (uniform)
+prior_pdf = prior.dist.pdf(p_range)
+ax1.plot(p_range, prior_pdf, label='Prior (Uniform)', color='blue', linewidth=2)
+
+# Plot posterior
+posterior_pdf = posterior.dist.pdf(p_range)
[…]
