+++ docs/examples/think-bayes-animals.md
@@ -1,0 +1,285 @@
+---
+comments: true
+---
+
+# Lions, Tigers and Bears: From Think Bayes Chapter 18
+
+This example ports the Lions, Tigers and Bears problem from Allen Downey's [Think Bayes Chapter 18](https://allendowney.github.io/ThinkBayes2/chap18.html#lions-and-tigers-and-bears) to demonstrate the Dirichlet-Multinomial conjugate relationship.
+
+The problem: estimating the prevalence of different animal species in a wildlife preserve, given observed counts. The Dirichlet distribution serves as the conjugate prior for the Multinomial likelihood, extending the Beta-Binomial concept to multiple categories.
+
+## Import modules
+
+Import the required distributions and functions:
+
+- `Dirichlet`: Prior and posterior distribution for species probabilities
+- `Multinomial`: Likelihood function for species counts
+- `DirichletMultinomial`: Predictive distribution for multinomial model
+- `multinomial_dirichlet`: Posterior update function
+- `multinomial_dirichlet_predictive`: Predictive distribution function
+
+```python
+from conjugate.distributions import Dirichlet, Multinomial, DirichletMultinomial
+from conjugate.models import multinomial_dirichlet, multinomial_dirichlet_predictive
+
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+```
+
+## Observed Data
+
+From Think Bayes: during a tour we observed 3 lions, 2 tigers, and 1 bear. We want to estimate the prevalence (probability) of each species.
+
+```python
+# Animal observation data
+observed_counts = np.array([3, 2, 1])  # [lions, tigers, bears]
+species_names = ['Lion', 'Tiger', 'Bear']
+total_observed = observed_counts.sum()
+
+print(f"Observed counts: {dict(zip(species_names, observed_counts))}")
+print(f"Total animals observed: {total_observed}")
+```
+
+## Bayesian Inference
+
+### Posterior Distribution
+
+Using the Dirichlet-Multinomial conjugate relationship, we update our prior with observed counts. The posterior is also a Dirichlet distribution:
+
+```python
+# Prior: Uniform Dirichlet(alpha=[1, 1, 1]) from Think Bayes
+prior = Dirichlet(alpha=np.array([1, 1, 1]))
+
+# Posterior update using conjugate relationship
+posterior = multinomial_dirichlet(x=observed_counts, prior=prior)
+
+print(f"Prior parameters: alpha={prior.alpha}")
+print(f"Posterior parameters: alpha={posterior.alpha}")
+print(f"Prior mean probabilities: {prior.mean()}")
+print(f"Posterior mean probabilities: {posterior.mean()}")
+
+# Display as DataFrame for clarity
+results_df = pd.DataFrame({
+    'Species': species_names,
+    'Observed': observed_counts,
+    'Prior Mean': prior.mean(),
+    'Posterior Mean': posterior.mean(),
+    'Observed Proportion': observed_counts / total_observed
+})
+print(results_df.round(3))
+```
+
+The posterior follows the simple conjugate update rule:
+`α_posterior = α_prior + observed_counts`
+
+### Predictive Distribution
+
+Get the predictive distribution for future observations:
+
+```python
+# Predictive distribution for next 10 animals
+n_future = 10
+posterior_predictive = multinomial_dirichlet_predictive(
+    n=n_future,
+    distribution=posterior
+)
+
[…]
