### 2026-02-10 b4b7754
Add figures to Think Bayes examples
- Generate comprehensive visualizations for all three examples
- World Cup: Prior/posterior comparison and predictive distributions
- Euro: Beta-Binomial analysis with credible intervals
- Animals: Dirichlet-Multinomial with parameter updates
- Update documentation to reference figure files
- High-quality PNG figures saved to docs/images/think-bayes/
- Figures clearly demonstrate conjugate prior advantages vs grid methods
