+++ docs/examples/think-bayes-world-cup.md
@@ -77,59 +77,9 @@ print(f"Predicted goals in next game: {posterior_predictive.mean():.2f}")
 
 Compare prior and posterior distributions to see how the data updates our beliefs about goal-scoring rates:
 
-```python
-# Create visualization
-fig, axes = plt.subplots(2, 1, figsize=(10, 8))
+![World Cup Problem](../images/think-bayes/world-cup.png)
 
-# Top plot: Prior vs Posterior distributions
-ax1 = axes[0]
-lam_range = np.linspace(0, 8, 200)
-
-# Plot prior
-prior_pdf = prior.dist.pdf(lam_range)
-ax1.plot(lam_range, prior_pdf, label='Prior', color='blue', linewidth=2)
-
-# Plot posterior
-posterior_pdf = posterior.dist.pdf(lam_range)
-ax1.plot(lam_range, posterior_pdf, label='Posterior', color='red', linewidth=2)
-
-# Add vertical lines for means
-ax1.axvline(prior.mean(), color='blue', linestyle='--', alpha=0.7, label=f'Prior mean: {prior.mean():.2f}')
-ax1.axvline(posterior.mean(), color='red', linestyle='--', alpha=0.7, label=f'Posterior mean: {posterior.mean():.2f}')
-ax1.axvline(x_total/n_games, color='black', linestyle=':', alpha=0.7, label=f'Observed rate: {x_total/n_games:.2f}')
-
-ax1.set_xlabel('Goal scoring rate (λ)')
-ax1.set_ylabel('Probability Density')
-ax1.set_title('Prior vs Posterior for Goal Scoring Rate')
-ax1.legend()
-ax1.grid(True, alpha=0.3)
-
-# Bottom plot: Predictive distribution
-ax2 = axes[1]
-goals_range = np.arange(0, 11)
-
-# Plot prior predictive
-prior_predictive = poisson_gamma_predictive(n=1, distribution=prior)
-prior_pmf = prior_predictive.dist.pmf(goals_range)
-ax2.bar(goals_range - 0.15, prior_pmf, width=0.3, label='Prior Predictive', color='blue', alpha=0.7)
-
-# Plot posterior predictive
-post_pmf = posterior_predictive.dist.pmf(goals_range)
-ax2.bar(goals_range + 0.15, post_pmf, width=0.3, label='Posterior Predictive', color='red', alpha=0.7)
-
-# Add observed data
-ax2.axvline(x_total, color='black', linestyle=':', linewidth=2, label=f'Observed: {x_total} goals')
-
-ax2.set_xlabel('Number of Goals')
-ax2.set_ylabel('Probability')
-ax2.set_title('Predictive Distribution for Next Game')
-ax2.set_xticks(goals_range)
-ax2.legend()
-ax2.grid(True, alpha=0.3)
-
-plt.tight_layout()
-plt.show()
-```
+*Figure: Prior vs Posterior distributions for goal-scoring rates, showing how observing 4 goals in 1 game updates our beliefs from a Gamma(1.4,1) prior to a Gamma(5.4,2) posterior.*
 
 ## Connection to Think Bayes
 
