+++ docs/examples/think-bayes-animals.md
@@ -112,126 +112,9 @@ for species, prob in zip(species_names, next_animal_probs):
 
 Create comprehensive visualizations of the analysis:
 
-```python
-# Create comprehensive visualization
-fig = plt.figure(figsize=(15, 10))
+![Lions, Tigers and Bears Problem](../images/think-bayes/animals.png)
 
-# 1. Prior vs Posterior mean probabilities
-ax1 = plt.subplot(2, 3, 1)
-x_pos = np.arange(len(species_names))
-width = 0.35
-
-ax1.bar(x_pos - width/2, prior.mean(), width, label='Prior Mean', color='blue', alpha=0.7)
-ax1.bar(x_pos + width/2, posterior.mean(), width, label='Posterior Mean', color='red', alpha=0.7)
-ax1.bar(x_pos + width*1.5, observed_counts/total_observed, width, label='Observed', color='green', alpha=0.7)
-
-ax1.set_xlabel('Species')
-ax1.set_ylabel('Probability')
-ax1.set_title('Prior vs Posterior Species Probabilities')
-ax1.set_xticks(x_pos)
-ax1.set_xticklabels(species_names)
-ax1.legend()
-ax1.grid(True, alpha=0.3)
-
-# 2. Dirichlet distributions (marginals)
-ax2 = plt.subplot(2, 3, 2)
-p_range = np.linspace(0, 1, 200)
-
-colors = ['blue', 'orange', 'green']
-for i, (species, color) in enumerate(zip(species_names, colors)):
-    # Get marginal beta distribution for this species
-    alpha_i = posterior.alpha[i]
-    alpha_sum = posterior.alpha.sum()
-
-    # Marginal beta distribution
-    from scipy.stats import beta as beta_dist
-    marginal_beta = beta_dist(alpha_i, alpha_sum - alpha_i)
-    pdf = marginal_beta.pdf(p_range)
-
-    ax2.plot(p_range, pdf, label=f'{species}', color=color, linewidth=2)
-    ax2.axvline(posterior.mean()[i], color=color, linestyle='--', alpha=0.7)
-
-ax2.set_xlabel('Probability')
-ax2.set_ylabel('Density')
-ax2.set_title('Marginal Posterior Distributions')
-ax2.legend()
-ax2.grid(True, alpha=0.3)
-
-# 3. Predictive distribution for next 10 animals
-ax3 = plt.subplot(2, 3, 3)
-
-# Sample from posterior predictive distribution
-n_samples = 1000
-predictive_samples = posterior_predictive.dist.rvs(size=n_samples, random_state=42)
-
-# Plot histogram of bear counts in samples
-bear_counts = predictive_samples[:, 2]  # Bear counts
-ax3.hist(bear_counts, bins=range(0, n_future+2), alpha=0.7, color='green', edgecolor='black')
-ax3.axvline(bear_counts.mean(), color='red', linestyle='--', linewidth=2,
-            label=f'Mean: {bear_counts.mean():.1f}')
-ax3.set_xlabel(f'Number of Bears in {n_future} Animals')
-ax3.set_ylabel('Frequency')
-ax3.set_title('Predictive Distribution: Future Observations')
-ax3.set_xticks(range(0, n_future+1))
-ax3.legend()
-ax3.grid(True, alpha=0.3)
-
-# 4. Prior vs Posterior parameter comparison
-ax4 = plt.subplot(2, 3, 4)
-
-ax4.scatter(prior.alpha, species_names, s=100, label='Prior α', color='blue', alpha=0.7)
-ax4.scatter(posterior.alpha, species_names, s=100, label='Posterior α', color='red', alpha=0.7)
-
-# Add lines to show updates
[…]
