### 2026-02-10 e0a0ed8
Fix import error in World Cup example
- Remove incorrect GammaPoisson import that doesn't exist
- Use correct imports: Gamma, Poisson with poisson_gamma_predictive
- All examples now have working imports
