### 2026-02-10 809c6c9
Fix predictive function names in Think Bayes examples
- Correct poisson_gamma_predictive function calls in World Cup example
- Correct binomial_beta_predictive function calls in Euro example
- Correct multinomial_dirichlet_predictive function calls in Animals example
- All examples now have working imports and function calls
