### 2026-02-13 williambdean/conjugate#305 @aneeshnaik
JOSS review: references not formatted correctly in paper
In the JOSS article proof, the various references are formatted incorrectly: there is no space between the prose text and the opening parenthesis of the citation (see screenshot below).
