### 2026-02-13 williambdean/conjugate#306 @aneeshnaik
JOSS review: distribution explorer not working as expected
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/9890.

This issue concerns the "Distribution Explorer" interface on the online docs (testing on Firefox 147.0.3 on an Apple machine).

Unless I'm misusing the tool, I believe it is not working as expected. In the screenshot below I have annotated 4 points:

1. The "Initializing..." icon does not go away.
2. These icons appear sometimes and then do not go away.
3. I was expecting the x axis of the plot (and the accompanying code snippet) to vary as I adjusted the `x min` and `x max` parameters above but this did not happen.
4. There is no functionality to vary the distribution parameters, which would be quite useful.
