### 2026-02-13 williambdean/conjugate#307 @aneeshnaik
JOSS review: a tabular overview of implemented conjugate distributions
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/9890.

I suppose users will typically approach `conjugate` with a specific use-case in hand, i.e., they want to implement a particular conjugate model. This was my situation when setting up my [biased coin test](https://gist.github.com/aneeshnaik/b5f3edcadfe7d939709f85ebc87d1b31). While I did eventually find all of the functionality I needed, and there were very relevant examples to help me, it was the case that it did take me a little while exploring the docs before I found the specific things I needed to implement my experiment.

In other words, it would be useful to have some resources in the docs to help users to quickly orient themselves around the package. My suggestion for this is an overview reference table listing all the implemented distributions/models, along with links to their documentation and any relevant examples from the examples page.

Something along the lines of:

| Conjugate Pair (likelihood / prior) | Model | Distribution | Examples |
|------|--------|-------|-------|
| Binomial / Beta | `binomial_beta` | `Beta` | Link 1, Link 2, Link 3 |

The `binomial_beta` and `Beta` entries here should be URLs linking to the relevant parts of the API reference, e.g. [`binomial_beta`](https://williambdean.github.io/conjugate/models/#conjugate.models.binomial_beta).

This table could go on the overview page, perhaps in place of the "Supported Models" section. Alternatively, it could serve as a page in its own right.
