### 2026-02-13 #9890 @aneeshnaik
I've completed my review of the submission. I have created 3 issues in the software repository:

- https://github.com/williambdean/conjugate/issues/305
- https://github.com/williambdean/conjugate/issues/306
- https://github.com/williambdean/conjugate/issues/307

Regarding the paper, I believe it is well written (although see my issue regarding formatting above) and successfully demonstrates that `conjugate` fills a gap in the statistical python ecosystem.

Regarding the software, I really like it and think it will be a good contribution to JOSS. As part of my review, I set up a simple experiment performing Bayesian inference with biased coin flips (experiment can be viewed at [this GitHub gist](https://gist.github.com/aneeshnaik/b5f3edcadfe7d939709f85ebc87d1b31)). The goals of this experiment were to test the functionality of the code, get a "feel" for the API, and test how helpful the documentation is at helping me implement my project.

Overall, this trial was a success: the code worked as expected, and the package had all the functionality I needed. In particular, the workflow for sequential Bayesian updating had a very pleasing and intuitive feel.

My most significant feedback from the trial is that I found it a little difficult to navigate the documentation to find exactly which functions / classes etc. were relevant for my use case. The documentation was all there (including some very relevant and helpful examples), but not intuitively easy to find. I have made a suggestion to address this issue in #307 above.
