### 2026-02-13 #9890 @williambdean
Thanks for that feedback @aneeshnaik . Very thorough and helpful feedback! That "Distribution Explorer" had a connection between data, model, and prior. However, that is broken from upstream issues at the moment. I will try to pin if possible but will go with a table solution in the mean time.

I will address the issues with the citation in the paper when I have the chance.

Thanks again
