### 2026-02-13 #9890 @danielskatz
@gcanasherrera - FYI, a command to editorialbot needs to the first thing in a new comment...
