+++ docs/quick-reference.md
@@ -1,0 +1,163 @@
+---
+hide:
+    - navigation
+comments: true
+---
+# Quick Reference
+
+A cheat sheet for conjugate model pairs to help you quickly find the right functions for your use case.
+
+## How to Use This Table
+
+1. Find your **likelihood** (the distribution of your data)
+2. Look up the corresponding **model function** from `conjugate.models`
+3. Check if there's a **helper function** to extract sufficient statistics from raw data
+4. See **examples** for practical usage patterns
+
+---
+
+## Discrete Likelihoods
+
+| Likelihood | Prior / Posterior | Model | Helper | Examples |
+|------------|-------------------|-------|--------|----------|
+| Binomial | [Beta] | [`binomial_beta`][binomial_beta] | [`binomial_beta_inputs`][binomial_beta_inputs] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
+| Bernoulli | [Beta] | [`bernoulli_beta`][bernoulli_beta] | [`bernoulli_beta_inputs`][bernoulli_beta_inputs] | [Raw Data Workflow] |
+| Negative Binomial | [Beta] | [`negative_binomial_beta`][negative_binomial_beta] | [`negative_binomial_beta_inputs`][negative_binomial_beta_inputs] | - |
+| Geometric | [Beta] | [`geometric_beta`][geometric_beta] | [`geometric_beta_inputs`][geometric_beta_inputs] | - |
+| Hypergeometric | [BetaBinomial] | [`hypergeometric_beta_binomial`][hypergeometric_beta_binomial] | [`hypergeometric_beta_binomial_inputs`][hypergeometric_beta_binomial_inputs] | - |
+| Categorical | [Dirichlet] | [`categorical_dirichlet`][categorical_dirichlet] | [`categorical_dirichlet_inputs`][categorical_dirichlet_inputs] | - |
+| Multinomial | [Dirichlet] | [`multinomial_dirichlet`][multinomial_dirichlet] | [`multinomial_dirichlet_inputs`][multinomial_dirichlet_inputs] | [Raw Data Workflow] |
+| Poisson | [Gamma] | [`poisson_gamma`][poisson_gamma] | [`poisson_gamma_inputs`][poisson_gamma_inputs] | [Bootstrap], [Limit], [Raw Data Workflow] |
+
+## Continuous Likelihoods
+
+| Likelihood | Prior / Posterior | Model | Helper | Examples |
+|------------|-------------------|-------|--------|----------|
+| Exponential | [Gamma] | [`exponential_gamma`][exponential_gamma] | - | [Thompson], [Raw Data Workflow] |
+| Gamma (known shape) | [Gamma] | [`gamma_known_shape`][gamma_known_shape] | [`gamma_known_shape_inputs`][gamma_known_shape_inputs] | - |
+| Gamma (known rate) | [GammaKnownRateProportional] | [`gamma_known_rate`][gamma_known_rate] | [`gamma_known_rate_inputs`][gamma_known_rate_inputs] | - |
+| Gamma | [GammaProportional] | [`gamma`][gamma] | [`gamma_inputs`][gamma_inputs] | - |
+| Inverse Gamma (known rate) | [Gamma] | [`inverse_gamma_known_rate`][inverse_gamma_known_rate] | [`inverse_gamma_known_rate_inputs`][inverse_gamma_known_rate_inputs] | - |
+| Normal (known variance) | [Normal] | [`normal_known_variance`][normal_known_variance] | [`normal_known_variance_inputs`][normal_known_variance_inputs] | [Shortest Path] |
+| Normal (known precision) | [Normal] | [`normal_known_precision`][normal_known_precision] | [`normal_known_precision_inputs`][normal_known_precision_inputs] | - |
[…]
