+++ docs/index.md
@@ -23,7 +23,7 @@ pip install conjugate-models
 
 - [Interactive Distribution Explorer](explorer.md) for exploring probability distributions with real-time parameter adjustment
 - **[Raw Data Workflow](examples/raw-data-workflow.md)** - Complete examples from raw observational data to posterior distributions with helper functions
-- **Data Input Helper Functions** - Extract sufficient statistics from raw observational data for all supported models
+- **[Data Input Helper Functions](helpers.md)** - Extract sufficient statistics from raw observational data for all supported models
 - [Connection to Scipy Distributions](examples/scipy-connection.md) with `dist` attribute
 - [Built in Plotting](examples/plotting.md) with `plot_pdf`, `plot_pmf`, and `plot_cdf` methods
 - [Vectorized Operations](examples/vectorized-inputs.md) for parameters and data
@@ -42,7 +42,11 @@ Many likelihoods are supported including
 - `Normal` (including linear regression)
 - and [many more](models.md)
 
+See the [Quick Reference](quick-reference.md) for a complete table of likelihood → prior/posterior mappings with links to model functions and helper functions.
+
 ## Basic Usage
+
+### Pattern 1: Working with Pre-processed Data
 
 1. Define prior distribution from `distributions` module
 1. Pass data and prior into model from `models` modules
@@ -52,9 +56,9 @@ Many likelihoods are supported including
 from conjugate.distributions import Beta, BetaBinomial
 from conjugate.models import binomial_beta, binomial_beta_predictive
 
-# Observed Data
-x = 4
-N = 10
+# Observed Data (sufficient statistics)
+x = 4  # successes
+N = 10 # trials
 
 # Analytics
 prior = Beta(1, 1)
@@ -65,6 +69,62 @@ posterior_predictive: BetaBinomial = binomial_beta_predictive(
     n=N, distribution=posterior
 )
 ```
+
+### Pattern 2: Working with Raw Observational Data
+
+For raw data, use **helper functions** from the `helpers` module to extract sufficient statistics:
+
+```python
+import numpy as np
+from conjugate.distributions import Beta
+from conjugate.models import binomial_beta
+from conjugate.helpers import bernoulli_beta_inputs
+
+# Raw observational data - individual trial outcomes
+raw_data = [1, 0, 1, 1, 0, 1, 0, 1, 1, 0]  # success/failure per trial
+
+# Extract sufficient statistics automatically
+inputs = bernoulli_beta_inputs(raw_data)
+print(inputs)  # {'x': 6, 'n': 10} - 6 successes in 10 trials
+
+# Use with conjugate model
+prior = Beta(1, 1)
+posterior = binomial_beta(prior=prior, **inputs)
+```
+
+#### Common Helper Function Patterns
+
+```python
+from conjugate.helpers import (
+    poisson_gamma_inputs,      # For count data
+    normal_known_variance_inputs,  # For continuous measurements
+    exponential_gamma_inputs,  # For time-between-events data
+    multinomial_dirichlet_inputs,  # For categorical data
+)
+
+# Count data (e.g., website visits per day)
+count_data = [5, 3, 8, 2, 6, 4, 7, 1, 9, 3]
+inputs = poisson_gamma_inputs(count_data)
+# Returns: {'x_total': sum(count_data), 'n': len(count_data)}
+
+# Continuous measurements with known variance
+measurements = [2.3, 1.9, 2.7, 2.1, 2.5]
+inputs = normal_known_variance_inputs(measurements)
+# Returns: {'x_total': sum(measurements), 'n': len(measurements)}
+# Note: variance must be passed separately to the model function
+
[…]
