+++ mkdocs.yml
@@ -81,6 +81,7 @@ plugins:
             docstring_style: google
 
 markdown_extensions:
+  - md_in_html
   - admonition
   - pymdownx.details
   - pymdownx.highlight:
