+++ docs/quick-reference.md
@@ -11,54 +11,72 @@ A cheat sheet for conjugate model pairs to help you quickly find the right funct
 
 1. Find your **likelihood** (the distribution of your data)
 2. Look up the corresponding **model function** from `conjugate.models`
-3. Check if there's a **helper function** to extract sufficient statistics from raw data
+3. Check the **Functions** column for:
+   - **Helper** functions to extract sufficient statistics from raw data
+   - **Predictive** functions for posterior predictive distributions
 4. See **examples** for practical usage patterns
+
+**Note:** Not all models have closed-form predictive distributions. When available, predictive functions follow the pattern `<model>_predictive`.
+
+---
+
+## Coverage
+
+This reference includes **26 conjugate model pairs** with:
+
+- ✅ **Model functions** for all 26 pairs (posterior updates)
+- ✅ **Helper functions** for 25 pairs (raw data → sufficient statistics)
+- ✅ **Predictive functions** for 18 pairs (posterior predictive distributions)
+
+Predictive functions are available for the most commonly used models, particularly those based on Beta, Gamma, and Normal priors.
+
+**Note:** For models without closed-form predictive distributions, see [Unsupported Distributions] for a guide on generating posterior predictive samples using sampling techniques.
 
 ---
 
 ## Discrete Likelihoods
 
-| Likelihood | Prior / Posterior | Model | Helper | Examples |
-|------------|-------------------|-------|--------|----------|
-| Binomial | [Beta] | [`binomial_beta`][binomial_beta] | [`binomial_beta_inputs`][binomial_beta_inputs] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
-| Bernoulli | [Beta] | [`bernoulli_beta`][bernoulli_beta] | [`bernoulli_beta_inputs`][bernoulli_beta_inputs] | [Raw Data Workflow] |
-| Negative Binomial | [Beta] | [`negative_binomial_beta`][negative_binomial_beta] | [`negative_binomial_beta_inputs`][negative_binomial_beta_inputs] | - |
-| Geometric | [Beta] | [`geometric_beta`][geometric_beta] | [`geometric_beta_inputs`][geometric_beta_inputs] | - |
-| Hypergeometric | [BetaBinomial] | [`hypergeometric_beta_binomial`][hypergeometric_beta_binomial] | [`hypergeometric_beta_binomial_inputs`][hypergeometric_beta_binomial_inputs] | - |
-| Categorical | [Dirichlet] | [`categorical_dirichlet`][categorical_dirichlet] | [`categorical_dirichlet_inputs`][categorical_dirichlet_inputs] | - |
-| Multinomial | [Dirichlet] | [`multinomial_dirichlet`][multinomial_dirichlet] | [`multinomial_dirichlet_inputs`][multinomial_dirichlet_inputs] | [Raw Data Workflow] |
-| Poisson | [Gamma] | [`poisson_gamma`][poisson_gamma] | [`poisson_gamma_inputs`][poisson_gamma_inputs] | [Bootstrap], [Limit], [Raw Data Workflow] |
+| Likelihood | Prior / Posterior | Model | Functions | Examples |
+|------------|-------------------|-------|-----------|----------|
+| Binomial | [Beta] | [`binomial_beta`][binomial_beta] | Helper: [`binomial_beta_inputs`][binomial_beta_inputs]<br>Predictive: [`binomial_beta_predictive`][binomial_beta_predictive] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
[…]
