+++ docs/quick-reference.md
@@ -38,45 +38,45 @@ Predictive functions are available for the most commonly used models, particular
 
 | Likelihood | Prior / Posterior | Model | Functions | Examples |
 |------------|-------------------|-------|-----------|----------|
-| Binomial | [Beta] | [`binomial_beta`][binomial_beta] | Helper: [`binomial_beta_inputs`][binomial_beta_inputs]<br>Predictive: [`binomial_beta_predictive`][binomial_beta_predictive] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
-| Bernoulli | [Beta] | [`bernoulli_beta`][bernoulli_beta] | Helper: [`bernoulli_beta_inputs`][bernoulli_beta_inputs]<br>Predictive: [`bernoulli_beta_predictive`][bernoulli_beta_predictive] | [Raw Data Workflow] |
-| Negative Binomial | [Beta] | [`negative_binomial_beta`][negative_binomial_beta] | Helper: [`negative_binomial_beta_inputs`][negative_binomial_beta_inputs]<br>Predictive: [`negative_binomial_beta_predictive`][negative_binomial_beta_predictive] | - |
-| Geometric | [Beta] | [`geometric_beta`][geometric_beta] | Helper: [`geometric_beta_inputs`][geometric_beta_inputs]<br>Predictive: [`geometric_beta_predictive`][geometric_beta_predictive] | - |
-| Hypergeometric | [BetaBinomial] | [`hypergeometric_beta_binomial`][hypergeometric_beta_binomial] | Helper: [`hypergeometric_beta_binomial_inputs`][hypergeometric_beta_binomial_inputs]<br>Predictive: - | - |
-| Categorical | [Dirichlet] | [`categorical_dirichlet`][categorical_dirichlet] | Helper: [`categorical_dirichlet_inputs`][categorical_dirichlet_inputs]<br>Predictive: [`categorical_dirichlet_predictive`][categorical_dirichlet_predictive] | - |
-| Multinomial | [Dirichlet] | [`multinomial_dirichlet`][multinomial_dirichlet] | Helper: [`multinomial_dirichlet_inputs`][multinomial_dirichlet_inputs]<br>Predictive: [`multinomial_dirichlet_predictive`][multinomial_dirichlet_predictive] | [Raw Data Workflow] |
-| Poisson | [Gamma] | [`poisson_gamma`][poisson_gamma] | Helper: [`poisson_gamma_inputs`][poisson_gamma_inputs]<br>Predictive: [`poisson_gamma_predictive`][poisson_gamma_predictive] | [Bootstrap], [Limit], [Raw Data Workflow] |
+| Binomial | [Beta][Beta-dist] | [`binomial_beta`][binomial_beta] | Helper: [`binomial_beta_inputs`][binomial_beta_inputs]<br>Predictive: [`binomial_beta_predictive`][binomial_beta_predictive] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
+| Bernoulli | [Beta][Beta-dist] | [`bernoulli_beta`][bernoulli_beta] | Helper: [`bernoulli_beta_inputs`][bernoulli_beta_inputs]<br>Predictive: [`bernoulli_beta_predictive`][bernoulli_beta_predictive] | [Raw Data Workflow] |
+| Negative Binomial | [Beta][Beta-dist] | [`negative_binomial_beta`][negative_binomial_beta] | Helper: [`negative_binomial_beta_inputs`][negative_binomial_beta_inputs]<br>Predictive: [`negative_binomial_beta_predictive`][negative_binomial_beta_predictive] | - |
[…]
