### 2026-02-15 9cebdeb
docs: SEO improvements & Quick Reference Table (#309)
* adding site description

* add quick reference

* validate changes

* update pypi info

* update pypi info

* adding to README and index

* add plugin

* add to quick refrence

* correct the referecnes
