+++ docs/quick-reference.md
@@ -1,0 +1,200 @@
+---
+hide:
+    - navigation
+comments: true
+---
+# Quick Reference
+
+A cheat sheet for conjugate model pairs to help you quickly find the right functions for your use case.
+
+## How to Use This Table
+
+1. Find your **likelihood** (the distribution of your data)
+2. Look up the corresponding **model function** from `conjugate.models`
+3. Check the **Functions** column for:
+   - **Helper** functions to extract sufficient statistics from raw data
+   - **Predictive** functions for posterior predictive distributions
+4. See **examples** for practical usage patterns
+
+**Note:** Not all models have closed-form predictive distributions. When available, predictive functions follow the pattern `<model>_predictive`.
+
+---
+
+## Coverage
+
+This reference includes **26 conjugate model pairs** with:
+
+- ✅ **Model functions** for all 26 pairs (posterior updates)
+- ✅ **Helper functions** for 25 pairs (raw data → sufficient statistics)
+- ✅ **Predictive functions** for 18 pairs (posterior predictive distributions)
+
+Predictive functions are available for the most commonly used models, particularly those based on Beta, Gamma, and Normal priors.
+
+**Note:** For models without closed-form predictive distributions, see [Unsupported Distributions] for a guide on generating posterior predictive samples using sampling techniques.
+
+---
+
+## Discrete Likelihoods
+
+| Likelihood | Prior / Posterior | Model | Functions | Examples |
+|------------|-------------------|-------|-----------|----------|
+| Binomial | [Beta][Beta-dist] | [`binomial_beta`][binomial_beta] | Helper: [`binomial_beta_inputs`][binomial_beta_inputs]<br>Predictive: [`binomial_beta_predictive`][binomial_beta_predictive] | [Binomial], [Bandit], [Bayesian Update], [Vectorized] |
+| Bernoulli | [Beta][Beta-dist] | [`bernoulli_beta`][bernoulli_beta] | Helper: [`bernoulli_beta_inputs`][bernoulli_beta_inputs]<br>Predictive: [`bernoulli_beta_predictive`][bernoulli_beta_predictive] | [Raw Data Workflow] |
+| Negative Binomial | [Beta][Beta-dist] | [`negative_binomial_beta`][negative_binomial_beta] | Helper: [`negative_binomial_beta_inputs`][negative_binomial_beta_inputs]<br>Predictive: [`negative_binomial_beta_predictive`][negative_binomial_beta_predictive] | - |
+| Geometric | [Beta][Beta-dist] | [`geometric_beta`][geometric_beta] | Helper: [`geometric_beta_inputs`][geometric_beta_inputs]<br>Predictive: [`geometric_beta_predictive`][geometric_beta_predictive] | - |
+| Hypergeometric | [BetaBinomial][BetaBinomial-dist] | [`hypergeometric_beta_binomial`][hypergeometric_beta_binomial] | Helper: [`hypergeometric_beta_binomial_inputs`][hypergeometric_beta_binomial_inputs]<br>Predictive: - | - |
+| Categorical | [Dirichlet][Dirichlet-dist] | [`categorical_dirichlet`][categorical_dirichlet] | Helper: [`categorical_dirichlet_inputs`][categorical_dirichlet_inputs]<br>Predictive: [`categorical_dirichlet_predictive`][categorical_dirichlet_predictive] | - |
[…]
