+++ conjugate/helpers.py
@@ -25,6 +25,8 @@ For PyTensor compatibility:
 import numpy as np
 from typing import Callable
 
+from conjugate._typing import NUMERIC
+
 # =============================================================================
 # Default Operations (NumPy-compatible, handle scalars)
 # =============================================================================
@@ -174,7 +176,7 @@ pytensor_ops = _PyTensorOpsProxy()
 
 
 def poisson_gamma_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -216,7 +218,7 @@ exponential_gamma_inputs = poisson_gamma_inputs
 
 
 def gamma_known_shape_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -249,8 +251,8 @@ def gamma_known_shape_inputs(
 
 
 def binomial_beta_inputs(
-    x,
-    n,
+    x: NUMERIC,
+    n: NUMERIC,
     *,
     sum_fn: Callable | None = None,
 ) -> dict:
@@ -274,7 +276,7 @@ def binomial_beta_inputs(
 
 
 def bernoulli_beta_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -316,8 +318,8 @@ def bernoulli_beta_inputs(
 
 
 def negative_binomial_beta_inputs(
-    x,
-    r,
+    x: NUMERIC,
+    r: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -358,7 +360,7 @@ def negative_binomial_beta_inputs(
 
 
 def geometric_beta_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -385,7 +387,7 @@ def geometric_beta_inputs(
 
 
 def hypergeometric_beta_binomial_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -417,7 +419,7 @@ def hypergeometric_beta_binomial_inputs(
 # =============================================================================
 
 
-def categorical_dirichlet_inputs(x) -> dict:
+def categorical_dirichlet_inputs(x: NUMERIC) -> dict:
     """Extract sufficient statistics for categorical_dirichlet model.
 
     Args:
@@ -429,7 +431,7 @@ def categorical_dirichlet_inputs(x) -> dict:
     return {"x": x}
 
 
-def multinomial_dirichlet_inputs(x) -> dict:
+def multinomial_dirichlet_inputs(x: NUMERIC) -> dict:
     """Extract sufficient statistics for multinomial_dirichlet model.
 
     Args:
@@ -448,7 +450,7 @@ def multinomial_dirichlet_inputs(x) -> dict:
 
 
 def normal_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -482,7 +484,7 @@ def normal_inputs(
 
 
 def normal_known_mean_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
     len_fn: Callable | None = None,
@@ -514,7 +516,7 @@ def normal_known_mean_inputs(
 
 
 def normal_known_variance_inputs(
-    x,
+    x: NUMERIC,
     *,
     sum_fn: Callable | None = None,
[…]
