### 2026-02-27 d159085
format links. closes #305
