### 2026-02-27 #9890 @williambdean
Yes, I have addressed 2 / 3. It is an upstream issue for the distribution explorer, noted here: https://github.com/williambdean/conjugate/issues/306
