### 2026-03-06 7a96d97
feat: auto-merge bot PRs when checks pass
- Add auto-merge job that triggers on bot PRs
- Wait for all status checks to pass (180 second timeout)
- Auto-merge with squash when checks succeed
- Fail gracefully if checks fail
- Respects branch protection rules

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
