### 2026-03-06 5049e59
feat: delete branch after auto-merge
Auto-delete bot PR branches after successful squash merge to keep
repository clean from remote branches.

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
