### 2026-03-06 4e24b49
fix: include bot-pr paths in test workflow triggers
Add uv.lock and .pre-commit-config.yaml to test workflow path filters
so tests run when Dependabot updates dependencies or pre-commit.ci
updates hooks. This ensures the auto-merge workflow doesn't hang
waiting for the Tests check.

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
