### 2026-03-06 e8e9a61
fix: trigger tests when tests.yml workflow changes
Include .github/workflows/tests.yml in test workflow path filters so
that changes to the test workflow definition itself trigger tests.

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
