### 2026-03-06 a018d5b
chore: auto-merge bot PRs when checks pass (#312)
* feat: auto-merge bot PRs when checks pass

- Add auto-merge job that triggers on bot PRs
- Wait for all status checks to pass (180 second timeout)
- Auto-merge with squash when checks succeed
- Fail gracefully if checks fail
- Respects branch protection rules

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

* feat: delete branch after auto-merge

Auto-delete bot PR branches after successful squash merge to keep
repository clean from remote branches.

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

*
[…]
