### 2026-03-06 c49708a
fix: replace polling loop with gh pr merge --auto
