+++ .github/workflows/bot-prs.yml
@@ -28,37 +28,7 @@ jobs:
     runs-on: ubuntu-latest
     if: github.event.pull_request.user.type == 'Bot'
     steps:
-      - uses: actions/checkout@v6
-      - name: Wait for checks to pass
-        run: |
-          number=${{ github.event.pull_request.number }}
-          max_attempts=180
-          attempt=0
-
-          while [ $attempt -lt $max_attempts ]; do
-            status=$(gh pr view $number --json statusCheckRollup --jq '.statusCheckRollup[].state' | sort | uniq)
-
-            if [[ "$status" == "SUCCESS" ]]; then
-              echo "All checks passed"
-              exit 0
-            elif [[ "$status" == *"FAILURE"* ]]; then
-              echo "Checks failed, will not auto-merge"
-              exit 1
-            fi
-
-            echo "Checks still running... attempt $((attempt + 1))/$max_attempts"
-            sleep 10
-            attempt=$((attempt + 1))
-          done
-
-          echo "Timeout waiting for checks"
-          exit 1
-        env:
-          GH_TOKEN: ${{ github.token }}
-
       - name: Auto-merge PR
-        run: |
-          number=${{ github.event.pull_request.number }}
-          gh pr merge $number --squash --auto --delete-branch
+        run: gh pr merge ${{ github.event.pull_request.number }} --squash --auto --delete-branch
         env:
           GH_TOKEN: ${{ github.token }}
