### 2026-03-06 d8fc2bd
fix: add GH_REPO env var to avoid needing git checkout (#314)
