+++ .github/workflows/bot-prs.yml
@@ -32,3 +32,4 @@ jobs:
         run: gh pr merge ${{ github.event.pull_request.number }} --squash --auto --delete-branch
         env:
           GH_TOKEN: ${{ github.token }}
+          GH_REPO: ${{ github.repository }}
