### 2026-03-11 #9890 @matt-graham
@gcanasherrera another nudge to ask about timelines for completing your review - if you are no longer able to complete the review please do say as we can proceed with just two reviewers.

@aneeshnaik - with regards to the remaining two unchecked items on your reviewer checklist, it looks like @williambdean addressed the issue (williambdean/conjugate#305) with references formatting in [this commit](https://github.com/williambdean/conjugate/commit/d159085d420dfaaf4a9514f78e031e9ffbbc676f). The functionality documentation item is also currently unticked - was this relating to williambdean/conjugate#307 as if so it appears this has now also been resolved?
