### 2026-03-13 #9890 @aneeshnaik
Apologies for the delay in returning to this! All the boxes in my checklist are ticked now, happy to recommend for publication. Thanks for addressing my comments @williambdean.

Regarding the problems with the Distribution Explorer, my suggestion would be to take it temporarily offline until the upstream issues are resolved, but definitely not worth holding up publication over this 🙂
