### 2026-03-13 #9890 @matt-graham
Hi @williambdean. I've read through the paper and have a few formatting and other minor suggested changes:

- [x] Line 6: 'API' acronym should be expanded on first usage e.g. _application programming interface_ (API).
- [x] Make use of _Title Case_ or _Sentence case_ in section headings consistent - currently _Statement of need_ is in sentence case while the remainder of the section headers are in title case.
- [x] Line 26: 'posterior computation is instantaneous regardless of data size' - I think this is a little misleading as updating the posterior parameters requires computing the relevant summary statistics of the data which will scale in cost with dataset size. Something like 'posterior computation cost is independent of dataset size once summary statistics have been computed' would be better I think.
- [x] Line 65: 'MCMC' acronym should be expanded on first usage, e.g. _Markov chain Monte Carlo_ (MCMC).
- [x] Lines 72, 86 and 93: scipy.stats should ideally be wrapped in backticks to render in fixed-width for consistency with references to other Python modules / subpackages such as `conjugate.models`.
- [x] Line 94: PDF acronym should be expanded in full to probability density function (does not appear to be used anywhere else).
- [x] Lines 147 - 150: list under _Paper Authoring_ paragraph should be either formatted as a bulleted list or an in-paragraph list for example using semi-colons as separators between the items.
- [x] Ideally package names such as NumPy, SciPy and Polars should be consistently capitalized - while these are capitalized as per the usage in the package documentation in the _Acknowledgements_ section, elsewhere in the paper these are generally rendered in all lowercase. Note pandas is an exception here as it explicitly refers to the package in all lowercase in its documentation.
- [x] Rather than the current `misc` type bibtex entry for Polars, it has a Zenodo record (https://zenodo.org/records/18983877) with accompanying DOI that would be a better to cite.
- [x] 'bayesian' and 'python' should be capitalized in ArviZ JOSS paper reference (`.bib` file entries need braces around these words to preserve capitalization).
