+++ .github/workflows/coverage.yml
@@ -28,6 +28,6 @@ jobs:
       - name: Run tests and collect coverage
         run: pytest
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v5.5.2
+        uses: codecov/codecov-action@v5.5.3
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
