+++ .github/workflows/tests.yml
@@ -49,7 +49,7 @@ jobs:
     - name: Test with tox
       run: tox
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v5.5.2
+      uses: codecov/codecov-action@v5.5.3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         name: ${{ matrix.python-version }}
