### 2026-04-09 williambdean/conjugate#305 @williambdean
This was fixed on the `paper` branch in commit d159085d420dfaaf4a9514f78e031e9ffbbc676f — spaces have been added between all prose text and `[@citation]` references throughout `paper/paper.md`. Closing.
