### 2026-04-09 williambdean/conjugate#324 @williambdean
Re-enable Distribution Explorer once upstream mkdocs-marimo issues are resolved
The Distribution Explorer has been temporarily disabled from the docs due to rendering issues in the upstream mkdocs-marimo plugin.

## Upstream Issues

- Spinner never resolves on load: https://github.com/marimo-team/mkdocs-marimo/issues/30
- Distribution parameter sliders not rendering: https://github.com/marimo-team/mkdocs-marimo/issues/45

## Context

This was raised by a reviewer during the JOSS review (#306) as the explorer was not functioning as expected.

## To Re-enable

Once both upstream issues above are resolved:

1. Revert docs/explorer.md to re-add the marimo_file embed
2. Remove the temporarily unavailable notices from docs/index.md and README.md
3. Verify the explorer loads and sliders work correctly
4. Close this issue and comment on #306
