### 2026-04-09 f1e5279
docs: temporarily disable Distribution Explorer until upstream mkdocs-marimo issues are resolved
Disables the interactive explorer embed in docs/explorer.md and adds
unavailability notices to docs/index.md and README.md, pointing to
tracking issue #324.

Upstream bugs:
- https://github.com/marimo-team/mkdocs-marimo/issues/30 (spinner)
- https://github.com/marimo-team/mkdocs-marimo/issues/45 (sliders)

Related: #306 (JOSS review)
