+++ docs/explorer.md
@@ -5,4 +5,9 @@ hide:
 ---
 # Distribution Explorer
 
-!marimo_file ./explorer.py
+> **Temporarily unavailable.** The interactive explorer is currently disabled due to rendering issues in the upstream `mkdocs-marimo` plugin:
+>
+> - Spinner never resolves on load ([mkdocs-marimo#30](https://github.com/marimo-team/mkdocs-marimo/issues/30))
+> - Distribution parameter sliders not rendering ([mkdocs-marimo#45](https://github.com/marimo-team/mkdocs-marimo/issues/45))
+>
+> The explorer will be re-enabled once these upstream issues are resolved. Progress is tracked in [issue #324](https://github.com/williambdean/conjugate/issues/324).
