### 2026-04-09 03c2e87
Apply remaining JOSS editor feedback to paper and bib
- Expand MCMC acronym on first use (Markov chain Monte Carlo)
- Wrap scipy.stats in backticks in all prose occurrences
- Expand PDF acronym on first use (probability density function)
- Fix section heading *Scipy Integration* -> *SciPy Integration*
- Add blank line before bullet list in Paper Authoring section
- Capitalize numpy->NumPy, polars->Polars, scipy->SciPy throughout prose
- Replace @misc{polars2024} with @software entry using Zenodo concept DOI
- Add BibTeX braces to ArviZ title to preserve capitalization of Bayesian/Python
