+++ paper/paper.bib
@@ -89,12 +89,12 @@
   url = {https://github.com/pandas-dev/pandas}
 }
 
-@misc{polars2024,
+@software{polars2024,
   author = {{Polars Contributors}},
   title = {Polars: Fast DataFrame Library},
   year = {2024},
-  url = {https://pola-rs.github.io/polars/},
-  note = {Accessed December 2025}
+  url = {https://github.com/pola-rs/polars},
+  doi = {10.5281/zenodo.7697217}
 }
 
 @article{gelman2020workflow,
@@ -126,7 +126,7 @@
   number = {33},
   pages = {1143},
   author = {Kumar, Ravin and Carroll, Colin and Hartikainen, Ari and Martin, Osvaldo},
-  title = {ArviZ a unified library for exploratory analysis of Bayesian models in Python},
+  title = {ArviZ a unified library for exploratory analysis of {Bayesian} models in {Python}},
   journal = {Journal of Open Source Software}
 }
 
