+++ paper/paper.md
@@ -19,9 +19,9 @@ bibliography: paper.bib
 
 # Summary
 
-`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic API and seamless integration with widely used Python data analysis libraries. It implements the conjugate likelihood-prior pairs cataloged in Fink's compendium and Wikipedia's conjugate prior table, making rigorous Bayesian updating, exploration, and visualization accessible for practitioners, educators, and researchers. Comprehensive documentation, interactive examples, and an online Distribution Explorer further support flexible application and learning.
+`conjugate-models` is a modern Python package for Bayesian conjugate inference that prioritizes a clean, idiomatic application programming interface (API) and seamless integration with widely used Python data analysis libraries. It implements the conjugate likelihood-prior pairs cataloged in Fink's compendium and Wikipedia's conjugate prior table, making rigorous Bayesian updating, exploration, and visualization accessible for practitioners, educators, and researchers. Comprehensive documentation, interactive examples, and an online Distribution Explorer further support flexible application and learning.
 
-# Statement of need
+# Statement of Need
 
 Bayesian inference with conjugate priors offers a tractable and interpretable
 approach to updating distributions in light of new evidence. While
@@ -30,9 +30,9 @@ significant cognitive and computational overhead for common conjugate models. Th
 package is designed to provide efficient, intuitive, and didactic support for
 Bayesian conjugate workflows across statistics, data science, and education,
 allowing direct use with array-like objects from libraries such as
-numpy [@harris2020array],
+NumPy [@harris2020array],
 pandas [@The_pandas_development_team_pandas-dev_pandas_Pandas], and
-polars [@polars2024]. The project also provides live interactive
+Polars [@polars2024]. The project also provides live interactive
 documentation and an [online Distribution Explorer](https://williambdean.github.io/conjugate/explorer) for real-time model investigation.
 
 A prior distribution is conjugate to a likelihood when the posterior remains in
@@ -40,8 +40,7 @@ the same distribution family after observing data [@raiffa1961applied].
 Conjugate priors provide closed-form posterior updates and posterior predictive
 distributions, eliminating the need for numerical
 integration [@fink1997compendium]. Because these updates are analytic
-rather than iterative, posterior computation is instantaneous regardless of
-data size—enabling real-time interactive exploration and rapid model
+rather than iterative, posterior computation cost is independent of dataset size once summary statistics have been computed—enabling real-time interactive exploration and rapid model
 iteration. `conjugate-models` implements the conjugate pairs cataloged in Fink's
 compendium [@fink1997compendium] and Wikipedia's conjugate prior table [@wikipedia_conjugate_prior].
 The complete list of supported models is maintained at
@@ -57,7 +56,7 @@ practitioners new to Bayesian methods.
 
 # Features & API Overview
 
-`conjugate-models` provides an intuitive, pipeable API compatible with numpy arrays [@harris2020array], pandas DataFrames/Series [@The_pandas_development_team_pandas-dev_pandas_Pandas], polars DataFrames [@polars2024] (for element-wise operations), and general numerical types. The package includes vectorized and indexable operations for batch and multi-arm inference, built-in plotting for posterior, prior, and predictive distributions, and connection to scipy distributions for interoperability [@virtanen2020scipy].
+`conjugate-models` provides an intuitive, pipeable API compatible with NumPy arrays [@harris2020array], pandas DataFrames/Series [@The_pandas_development_team_pandas-dev_pandas_Pandas], Polars DataFrames [@polars2024] (for element-wise operations), and general numerical types. The package includes vectorized and indexable operations for batch and multi-arm inference, built-in plotting for posterior, prior, and predictive distributions, and connection to SciPy distributions for interoperability [@virtanen2020scipy].
 
[…]
