### 2026-04-09 #9890 @williambdean
@matt-graham — the paper and software are ready for final acceptance checks:

- **Software version**: v0.14.2
- **Zenodo archive**: https://doi.org/10.5281/zenodo.19488126
- **GitHub release**: https://github.com/williambdean/conjugate/releases/tag/v0.14.2
- **Paper branch**: https://github.com/williambdean/conjugate/tree/paper

All editor feedback from your 2026-03-13 comment has been addressed in the paper branch, including: acronym expansions (API, MCMC, PDF), `scipy.stats` backtick formatting, capitalization of NumPy/SciPy/Polars, bullet list blank line fix, `*SciPy Integration*` heading correction, Polars `@software` bib entry with Zenodo DOI, and ArviZ title bracing.
