### 2026-04-10 #9890 @matt-graham
Thanks @williambdean for making those edits - that all looks good now. For the Zenodo archive, the title and author list in the metadata need to match those in the paper. It looks like you have used the GitHub Zenodo integration which auto populates title from repository name and uses contributor list for authors. You should be able to manually edit the metadata for the existing Zenodo archive without making a new release. Once you've done that let me know and I'll do final actions.
