### 2026-04-13 #9890 @matt-graham
Hi @williambdean. Thanks for updating the Zenodo metadata. Can the backticks around conjugate-models on the title on Zenodo also be removed as we require the title to exactly match (and we generally advise against formatting syntax in titles as it can issues when rendering the title).
