### 2026-04-20 #9890 @matt-graham
Hi @williambdean - just wanting to check if you saw [my message above ](https://github.com/openjournals/joss-reviews/issues/9890#issuecomment-4234711271) with regards to removing backticks in title metadata for Zenodo archive. Once that's done I can proceed with the next steps.
