### 2026-04-21 #9890 @williambdean
@matt-graham I've removed the backticks from the title in the Zenodo metadata. The title now reads: `conjugate-models: Conjugate Models in Python`. Let me know if there's anything else needed!
