### 2026-04-25 #9890 @crvernon
@williambdean - this submission is in great shape.  I just need you to address the following before I accept:

In your paper:

- [x] LINE 19:  Please use the correct citation for Pandas that includes the date.  See the recommended citations [here](https://pandas.pydata.org/about/citing.html).
- [x] LINES 20-21:  Please ensure that the link here is pointing to a page that is functioning correctly.

Let me know when these changes have been addressed.  Thanks.
