### 2026-04-29 94136a2
Address AEiC feedback: update pandas citation and remove Distribution Explorer references
- Update pandas bib entry to v3.0.2 with correct DOI, month, year, publisher,
  and version fields per pandas recommended citation format
- Remove all four references to the Distribution Explorer from paper, as it is
  temporarily disabled (tracked in issue #324)
