+++ paper/paper.bib
@@ -82,11 +82,14 @@
 }
 
 @software{The_pandas_development_team_pandas-dev_pandas_Pandas,
-  author = {{The pandas development team}},
-  doi = {10.5281/zenodo.3509134},
-  license = {BSD-3-Clause},
-  title = {{pandas-dev/pandas: Pandas}},
-  url = {https://github.com/pandas-dev/pandas}
+  author    = {{The pandas development team}},
+  title     = {{pandas-dev/pandas: Pandas}},
+  month     = mar,
+  year      = 2026,
+  publisher = {Zenodo},
+  version   = {v3.0.2},
+  doi       = {10.5281/zenodo.19340003},
+  url       = {https://doi.org/10.5281/zenodo.19340003}
 }
 
 @software{polars2024,
