### 2026-04-29 #9890 @williambdean
@crvernon both items have been addressed on the paper branch:

1. Updated the pandas citation to the latest recommended entry (v3.0.2, DOI: 10.5281/zenodo.19340003) per the pandas citation guidelines.
2. Removed all references to the Distribution Explorer, which is temporarily disabled pending an upstream fix (tracked in https://github.com/williambdean/conjugate/issues/324).

Let me know if anything else is needed!
